pyproject.toml
[tool.poetry]
name = "spaced_repetition_bot"
version = "0.1.0"
description = "Bot for spaced repetition"
authors = ["Hiyorimi"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.11.2"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
